分类联动required属性；默认post表单双栏布局；优化字段必选JS检测；优化Action php语法

+++ includes/action/wnd-do-pay.php
@@ -12,8 +12,8 @@ use Wnd\Model\Wnd_Payment;
 class Wnd_Do_Pay extends Wnd_Action_Ajax {
 
 	public static function execute(): array{
-		$post_id         = $_POST['post_id'] ?? 0;
-		$total_amount    = $_POST['total_amount'] ?? 0;
+		$post_id         = (int) ($_POST['post_id'] ?? 0);
+		$total_amount    = (float) ($_POST['total_amount'] ?? 0);
 		$payment_gateway = $_POST['payment_gateway'] ?? '';
 
 		if (!$payment_gateway) {
