+++ includes/action/wnd-update-post.php
@@ -15,7 +15,7 @@ class Wnd_Update_Post extends Wnd_Action_Ajax {
 	 */
 	public static function execute($post_id = 0): array{
 		// 获取被编辑Post ID
-		$_POST['_post_ID'] = $post_id ?: $_POST['_post_ID'];
+		$_POST['_post_ID'] = (int) ($post_id ?: $_POST['_post_ID']);
 		if ($_POST['_post_ID']) {
 			return ['status' => 0, 'msg' => __('ID无效', 'wnd')];
 		}
